requirements.txt
flask
numpy
pillow
scikit-learn
joblib
